requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
redis==4.3.4
psycopg2-binary==2.9.3
Flask-Mail==0.9.1

